fix: move logic from the controller to the service

+++ src/main/java/cat/itacademy/s04/t02/n01/S04T02N01/service/FruitService.java
@@ -5,6 +5,10 @@ import cat.itacademy.s04.t02.n01.S04T02N01.model.exceptions.ResourceNotFoundExce
 import cat.itacademy.s04.t02.n01.S04T02N01.repository.FruitRepository;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
+import java.util.stream.Collectors;
+import java.util.stream.StreamSupport;
+
 @Service
 public class FruitService {
 
@@ -18,11 +22,18 @@ public class FruitService {
         return fruitRepository.save(fruit);
     }
 
-    public Fruit updateFruit(Fruit fruit) {
+    public Fruit updateFruit(int id, Fruit fruit) {
+        Fruit existingFruit = fruitRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Fruit with ID " + id + " not found"));
+
+        fruit.setId(existingFruit.getId());
         return fruitRepository.save(fruit);
     }
 
     public void deleteFruit(int id) {
+        if (!fruitRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Fruit with ID " + id + " not found");
+        }
         fruitRepository.deleteById(id);
     }
 
@@ -31,7 +42,9 @@ public class FruitService {
                 .orElseThrow(() -> new ResourceNotFoundException("Fruit with ID " + id + " not found"));
     }
 
-    public Iterable<Fruit> getAllFruit() {
-        return fruitRepository.findAll();
+    public List<Fruit> getAllFruit() {
+        return StreamSupport.stream(fruitRepository.findAll().spliterator(), false)
+                .collect(Collectors.toList());
     }
+
 }
